bugfix: order of nodes in DancingLinks
Updated addRow method so that the 'right' reference of the nodes match
those of the column header nodes.

For example, suppose we have column headers
header -> A -> B -> C -> D -> header -> ...
where '->' represents the 'right' references (e.g. A.right == B).
If, for example, a row has bits set at A, B and D, it should looks like
A -> B -> D -> A -> ...
However, previously it would look like this instead
A -> D -> B -> A -> ...