parallelize DancingLinks
Allow DancingLinks to find solutions in parallel. The number of threads
and search depth is specified as an argument to the getSolutions method.
The depth is the depth to search for partial solutions. The partial
solutions will then be used as a starting point for separate tasks which
can be run on multiple threads.